+++ auto-setup/functions/dockerCoreRegistryDeleteRepositoryTags.sh
@@ -1,32 +1,35 @@
-dockerCoreRegistryDeleteRepositoryTags() { 
+dockerCoreRegistryDeleteRepositoryTags() {
 
-  # Call common function to execute common function start commands, such as setting verbose output etc
-  functionStart
+  local imagePath=${1:-}
+  local dockerRegistryPassword=$(getPassword "docker-registry-${namespace}-password" "docker-registry")
 
-  local imagePath=${1}
-  local dockerRegistryPassword=$(getPassword "docker-registry-${baseUser}-password" "docker-registry")
+  if [[ -n ${imagePath} ]]; then
 
-  # Get repository tags
-  local imageTags=$(dockerCoreRegistryGetImageTags "${imagePath}")
+    # Get repository tags
+    local imageTags=$(dockerCoreRegistryGetImageTags "${imagePath}")
 
-  # Delete image tags from registry
-  for imageTag in ${imageTags}
-  do
-    # Get image tag sha
-    local imageTagSha=$(dockerCoreRegistryGetImageShaDigest "${imagePath}" "${imageTag}")
+    # Delete image tags from registry
+    for imageTag in ${imageTags}
+    do
+      # Get image tag sha
+      local imageTagSha=$(dockerCoreRegistryGetImageShaDigest "${imagePath}" "${imageTag}")
 
-    # Delete image with sha reference
-    curl -u ${baseUser}:${dockerRegistryPassword} \
-     -X DELETE https://docker-registry.${baseDomain}/v2/${imagePath}/manifests/${imageTagSha}
-  done
+      if [[ -n ${imageTagSha} ]]; then
+        # Delete image with sha reference
+        curl -u ${namespace}:${dockerRegistryPassword} \
+        -X DELETE https://docker-registry.${baseDomain}/v2/${imagePath}/manifests/${imageTagSha}
+      fi
 
-  # Prune Docker Registry
-  kubectl exec $(kubectl get pod -n docker-registry -o name) -c docker-registry -n docker-registry -- bin/registry garbage-collect /etc/docker/registry/config.yml --delete-untagged=true
+    done
 
-  # Clean up local file system as well
-  docker system prune --force
+    if [[ -n ${imageTags} ]]; then
+      # Prune Docker Registry
+      kubectl exec $(kubectl get pod -n docker-registry -o name) -c docker-registry -n docker-registry -- bin/registry garbage-collect /etc/docker/registry/config.yml --delete-untagged=true || log_warn "Nothing to prune as no \"${imagePath}\" images published into docker registry - \"https://docker-registry.${baseDomain}\""
+    fi
 
-  # Call common function to execute common function start commands, such as unsetting verbose output etc
-  functionEnd
+    # Clean up local file system as well
+    docker system prune --force
+
+  fi
 
 }
